move `redux` folder to `src`
