+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { RestaurantsPage } from "src/pages/restaurants-page/component";
 import "./scss/styles.scss";
 import { Provider } from "react-redux";
-import store from "src/components/redux";
+import store from "src/redux";
 
 export const App: React.FC = () => {
   return (
